saga_log/requirements.txt
flask==3.1.0
flask_sqlalchemy==3.1.1
psycopg2-binary==2.9.10
marshmallow_sqlalchemy==1.4.1
pulsar-client==3.6.1
pulsar-client[avro]==3.6.1